No longer compile native bits (#8)
Upstream JNA is currently supporting older versions of glibc. And
getting the build environment set up to recompile the native libraries
is now very challenging due to Centos 6 going end of life. Instead of
jumping through hoops, we will revert to relying on upstream
JNA. However, we will retain repackaging the JAR to only include native
libraries for platforms that we support.